Change Noria prepare update statement id to u32
There was no discernable reason for why this was set to u64. This commit
changes this to u32 to be consistent with the statement ids set for all
other perpare types.

Reviewed-on: https://gerrit.readyset.name/c/readyset/+/646
Tested-by: Buildkite CI

+++ noria-client/src/backend.rs
@@ -419,10 +419,8 @@ where
                     .insert(self.prepared_count, PreparedStatement::Noria(*statement_id));
             }
             PrepareResult::Noria(Update { statement_id, .. }) => {
-                self.prepared_statements.insert(
-                    self.prepared_count,
-                    PreparedStatement::Noria(*statement_id as u32),
-                );
+                self.prepared_statements
+                    .insert(self.prepared_count, PreparedStatement::Noria(*statement_id));
             }
             PrepareResult::Upstream(UpstreamPrepare { statement_id, .. }) => {
                 self.prepared_statements.insert(self.prepared_count, {
